+++ pnet-data-api-java/src/main/java/pnet/data/api/application/ApplicationItemDTO.java
@@ -21,6 +21,7 @@ import java.util.Collection;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import io.swagger.v3.oas.annotations.media.Schema;
+import pnet.data.api.util.WithDescription;
 import pnet.data.api.util.WithLabel;
 import pnet.data.api.util.WithLastUpdate;
 import pnet.data.api.util.WithMatchcode;
@@ -32,7 +33,8 @@ import pnet.data.api.util.WithScore;
  * @author ham
  */
 @Schema(description = "Holds basic information about an application")
-public class ApplicationItemDTO implements WithMatchcode, WithLabel, WithLastUpdate, WithScore, Serializable
+public class ApplicationItemDTO
+    implements WithMatchcode, WithLabel, WithDescription, WithLastUpdate, WithScore, Serializable
 {
     private static final long serialVersionUID = 1943888464506455363L;
 
@@ -41,6 +43,9 @@ public class ApplicationItemDTO implements WithMatchcode, WithLabel, WithLastUpd
 
     @Schema(description = "The label of the application in the requested language.")
     private final String label;
+
+    @Schema(description = "The description of the application in the requested language.")
+    private final String description;
 
     @Schema(description = "The matchcodes of all scopes this application is permitted to use.")
     private final Collection<String> scopeMatchcodes;
@@ -52,11 +57,13 @@ public class ApplicationItemDTO implements WithMatchcode, WithLabel, WithLastUpd
     private final double score;
 
     public ApplicationItemDTO(@JsonProperty("matchcode") String matchcode, @JsonProperty("label") String label,
+        @JsonProperty("description") String description,
         @JsonProperty("scopeMatchcodes") Collection<String> scopeMatchcodes,
         @JsonProperty("lastUpdate") LocalDateTime lastUpdate, @JsonProperty("score") double score)
     {
         super();
         this.matchcode = matchcode;
+        this.description = description;
         this.label = label;
         this.scopeMatchcodes = scopeMatchcodes;
         this.lastUpdate = lastUpdate;
@@ -73,6 +80,12 @@ public class ApplicationItemDTO implements WithMatchcode, WithLabel, WithLastUpd
     public String getLabel()
     {
         return label;
+    }
+
+    @Override
+    public String getDescription()
+    {
+        return description;
     }
 
     public Collection<String> getScopeMatchcodes()
@@ -96,7 +109,8 @@ public class ApplicationItemDTO implements WithMatchcode, WithLabel, WithLastUpd
     public String toString()
     {
         return String
-            .format("ApplicationItemDTO [matchcode=%s, label=%s, scopeMatchcodes=%s, lastUpdate=%s, score=%s]",
-                matchcode, label, scopeMatchcodes, lastUpdate, score);
+            .format(
+                "ApplicationItemDTO [matchcode=%s, label=%s, description=%s, scopeMatchcodes=%s, lastUpdate=%s, score=%s]",
+                matchcode, label, description, scopeMatchcodes, lastUpdate, score);
     }
 }