+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-pnet-data-api 2.1.0
+pnet-data-api 2.2.0
+===================
+
+* Add description to application item.
+
+pnet-data-api 2.1.1
 ====================
 
 * Add replace and remove attribute methods to RestCall.
@@ -34,6 +39,11 @@ pnet-data-api 2.0.0
   older Spring Boot version (since the Data API does not depend on it, nor does is use any Jakarta packages). You still
   have to use up-to-date libraries like Jackson 2.14 or greater, Spring-Web 6.0 or greater (if using the Spring
   connector) and Java 17 or greater.
+
+pnet-data-api 1.23.0
+====================
+
+* Add description to application item.
 
 pnet-data-api 1.22.1
 ====================